+++ src/types.ts
@@ -1,0 +1,5 @@
+export interface GregorianDate {
+    year: number,
+    month: number,
+    day: number,
+}
